requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
